refactor : 공지사항 컴포넌트 리팩터링
